+++ js/scripts.ts
@@ -1,5 +1,4 @@
 // import im from 'include-media-export/include-media.js';
-import 'instant.page';
 
 ( function() {
 	// Remove the 'no-js' <body> class